requirements.txt
numpy
pandas
scikit-learn
tensorflow-cpu
selenium
joblib
igramscraper
requests
djangorestframework
gunicorn
django-heroku
django-apscheduler